Bugfix: TimedScanTrack modified spintime wrongly rarely.
The TimedScanTrack method tries to find a more precise spintime by
finding the same sector header at the beginning of first and second
revolution and the difference of their time is the new spintime.
The bug appears when the headers are weak (not always detected) and
the first header detected in the second revolution does not have a
matching header in the first revolution. In this case the method
assumed that the first sector in first revolution can be used but this
is wrong. Thus the spintime was corrected wrongly which cased unwanted
sideaffects.
Solution: the spintime is adjusted only if the same sector header pair
is detected at the beginning of first and second revolution.