Bugfix: TimedScanTrack normalised sector offsets wrongly rarely.
The TimedScanTrack method normalises the sector offsets because
rarely a sector can be detected in the second revolution that is
after the spintime. However sector offsets must be less than spintime
Earlier the method simply subtracted the spintime from sector offset
which was not enough when a weak track does not return any headers for
the first two tries but returns headers for the third try. In this
case the offset is greater than spintime * 2 thus subtracting spintime
once is not enough.
Solution: the offset is calculated as offset modulo spintime instead of
offset - spintime.