Upgrade: FAMs are replaced with methods because they FAMs are C features.
FAM is flexible array member which is an emtpy array at the end of a
structure. It is a C feature and the project is now C++ thus an upgrade
was necessary. It is replaced with two methods. The first returns the
address of array, the second returns the nth element of the array.