Added a KdCounter integer so KdPrint methods can display it as index.
Debugging a kernel driver on a local computer is quite limited. The
driver can print kernel debug messages by KdPrint which can be watched
in a WinDbg or such. However the messages are usually wrapped around
because the KdPrint uses a ring buffer. Thus when displaying these
messages a block of messages can overtake another block of messages
although it should not happen according to the time when the messages
were created. To help this situation the new KdCounter variable can be
sent to KdPrint and it should be incremented at the same time. This way
the order of messages becomes clear.