Adding Get Multi Track Time feature.
Based on the idea of Get Track Time feature, the new Multi Track Time
feature is added. Its purpose is similar to the single
version but it operates with measuring the spintime multiple times.
This way a more precise spintime can be achieved. More exactly the
average of spintimes can be achieved.